LPS-61952 Move com.liferay.portlet.imagegallerydisplay.display.context to com.liferay.image.gallery.display.kernel.display.context
