+++ portal-service/src/com/liferay/image/gallery/display/kernel/display/context/BaseIGDisplayContext.java
@@ -12,12 +12,26 @@
  * details.
  */
 
-package com.liferay.portlet.imagegallerydisplay.display.context;
+package com.liferay.image.gallery.display.kernel.display.context;
 
-import com.liferay.portlet.documentlibrary.display.context.DLUIItemKeys;
+import com.liferay.portal.kernel.display.context.BaseDisplayContext;
+
+import java.util.UUID;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 /**
  * @author Iván Zaera
  */
-public class IGUIItemKeys extends DLUIItemKeys {
+public abstract class BaseIGDisplayContext<T extends IGDisplayContext>
+	extends BaseDisplayContext<T> implements IGDisplayContext {
+
+	public BaseIGDisplayContext(
+		UUID uuid, T parentIGDisplayContext, HttpServletRequest request,
+		HttpServletResponse response) {
+
+		super(uuid, parentIGDisplayContext, request, response);
+	}
+
 }
